refactor(core): :recycle: rename branding-preference-text file to get-branding-preference-text
