+++ packages/core/src/api/get-branding-preference-text.ts
