+++ packages/core/src/api/public-api.ts
@@ -20,4 +20,4 @@ export {default as authorize} from './authorize';
 export {default as authenticate} from './authenticate';
 export {default as getBrandingPreference} from './branding-preference';
 export {default as getProfileInformation} from './profile';
-export {default as getBrandingPreferenceText} from './branding-preference-text';
+export {default as getBrandingPreferenceText} from './get-branding-preference-text';
